Use make to make artifacts instead of phonies
Make is better when used to make artifacts so we can use it here to
actually make the artifacts instead of relying on a single target to
make everything. Changes the travis yaml and Dockerfile to utilize
the targets as well.

Also gives the option of having `bin/inlets` produce an artifact that
could be native to your system (instead of having a specific target to
produce a darwin specific binary).

There's also some niceties thrown in like having the some basic make
targets like `clean`, `build`, `install`, and `uninstall`.

Signed-off-by: Eli Uriegas <seemethere101@gmail.com>