Switch to writable test in get.sh
The evaluation prior to moving the binary in get.sh was checking for the root user.  However, if the current user has sufficient privileges to write to the target location then there would be no reason not to attempt the move as the current user.

This change moves the if statement away from testing for the root user and instead checks for writability for the current location.

Signed-off-by: Richard Gee <richard@technologee.co.uk>