Enable strict filtering of upstream hosts
There was an issue reported by a user about the allow code
which meant any host could be set by the server in a dial
request without any validation. This behaviour was as designed,
however some users may want a strict list of upstream hosts
to be the only ones that can be contacted.

Tested by hacking the server code to change the host being
dialed to one not specified by the client. This resulted in
a disconnect and denial of access.

Update README for strict forwarding flag.

Optional feature: off by default.

Fixes: #165

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>