Update k8s.io/apimachinery
This dependency can be updated, however it brings in a number
of unexecpted transitive dependencies making the resulting
output larger.

Under a basic load test proxying SimpleHTTPServer, hey was
hanging for a few seconds towards the end of a test run
which may be related to the beta version of Go running on
my Mac M1.

It may make sense to dig further into what's changed in the
proxy package that may have affected connections.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>